Removes old scripts
