+++ lib/index.js
@@ -4,14 +4,14 @@ const { parse } = require('@babel/parser');
 const traverse = require('@babel/traverse').default;
 const t = require('@babel/types');
 
-const ASTToCsharp = require('./utils/ast-to-csharp');
-const createNewDefinitions = require('./transforms/create-new-definitions');
+const createDefinitions = require('./transforms/create-definitions');
 const expandReferences = require('./transforms/expand-references');
 const getComponentName = require('./utils/get-component-name');
 const getMeta = require('./utils/get-meta');
 const getPropTypes = require('./utils/get-prop-types');
 const getPropTypesIdentifierName = require('./utils/get-prop-types-identifier-name');
 const stripPropTypesIdentifier = require('./utils/strip-prop-types-identifier');
+const stringify = require('./stringify');
 const transformPropTypes = require('./transforms/transform-prop-types');
 const validateProps = require('./utils/validate-props');
 
@@ -73,14 +73,10 @@ module.exports = function({ baseClass, indent, namespace, sourceCode }) {
 
   const propTypes = parseTree(syntaxTree);
 
-  const newTransform = require('./transforms/transform-prop-types2');
-
-  const transformedPropTypes = newTransform(propTypes, propTypesMeta);
+  const transformedPropTypes = transformPropTypes(propTypes, propTypesMeta);
 
   // Strip client-only types, apply meta types and strip 'PropTypes' prefix from types
   // transformPropTypes({ propTypesMeta, syntaxTree });
-
-  const createDefinitions = require('./transforms/create-definitions');
 
   const definitions = createDefinitions(transformedPropTypes, componentName);
 
@@ -104,9 +100,7 @@ module.exports = function({ baseClass, indent, namespace, sourceCode }) {
   // };
   // EnumArray = ['value-1', 'value-2'];
 
-  const classesToString = require('./stringify/classes-to-string');
-
-  const code = classesToString({
+  const code = stringify({
     baseClass,
     componentName,
     definitions,
