+++ lib/stringify/capitalize.js
