+++ lib/stringify/generator.js
@@ -1,4 +1,4 @@
-const capitalize = require('../utils/capitalize');
+const capitalize = require('./capitalize');
 
 const enumProperties = require('./enum-properties');
