+++ lib/stringify/index.js
