+++ lib/stringify/unknown-to-pascal.js
