+++ lib/transforms/create-definitions.js
@@ -1,6 +1,3 @@
-const capitalize = require('../utils/capitalize');
-const isIdentifierArray = require('../utils/is-identifier-array');
-
 // This function expects the component name as string, and an object representing the propTypes of a component.
 // It returns an array of objects representing the classes to be generated
 //  - Every prop of type 'oneOf' and 'shape' will get a class definition
