+++ lib/transforms/transform-prop-types.js
@@ -1,6 +1,3 @@
-const traverse = require('@babel/traverse').default;
-const t = require('@babel/types');
-
 const getMetaNodeForProp = require('../utils/get-meta-node-for-prop');
 
 const typesToStrip = ['element', 'func', 'instanceOf', 'node'];
@@ -11,92 +8,62 @@ const illegalTypes = {
   oneOfType: `'PropTypes.oneOfType' is not yet supported`
 };
 
-// This function mutates the provided syntaxTree, doing the following (in this order):
+// This function mutates the provided propTypes tree, doing the following:
 //  - Excludes props that match typesToStrip or have an 'exclude' meta type
 //  - Replaces PropType definitions with meta types if provided
 //  - Validates type against illegalTypes
-//  - Validates References to other components in 'arrayOf'
 
-// This is a lot of different things in a single function, but it is necessary in order to avoid excessive traversal of the syntax tree.
+// Expand references from 'shape'
+// shape(Component) -> Component
+const expandShapeReference = propType => {
+  if (propType.type === 'shape' && propType.argument) {
+    return expandShapeReference(propType.argument);
+  }
 
-const maybeRequiredNode = (node, isRequired) => {
-  return isRequired
-    ? t.memberExpression(node, t.identifier('isRequired'))
-    : node;
+  return propType;
 };
 
-module.exports = function({ propTypesMeta, syntaxTree }) {
-  traverse(syntaxTree, {
-    MemberExpression(path) {
-      // Replace 'Component.propTypes' with 'Component'
-      if (path.get('property').isIdentifier({ name: 'propTypes' })) {
-        path.replaceWith(path.node.object);
-      }
-    },
-    ObjectProperty(path) {
-      const valuePath = path.get('value');
-      const value = path.node.value;
-      const propName = path.node.key.name;
-      const isFunction = t.isCallExpression(value);
-      const typeNode = isFunction ? value.callee : value;
-      const isMember = t.isMemberExpression(typeNode);
-      const typeName = isMember ? typeNode.object.name : typeNode.name;
-      const meta = getMetaNodeForProp({ path, propTypesMeta });
-      const isRequired =
-        isMember && t.isIdentifier(typeNode.property, { name: 'isRequired' });
+const transformPropTypes = (propTypes, propTypesMeta) => {
+  // 'propTypes' is a string when 'A.propTypes = B.propTypes;'
+  if (typeof propTypes === 'string') {
+    return propTypes;
+  }
 
-      // Strip if type matches typesToStrip or has 'exclude' meta type
-      if (
-        typesToStrip.includes(typeName) ||
-        t.isIdentifier(meta, { name: 'exclude' })
-      ) {
-        return path.remove();
-      }
+  return Object.entries(propTypes).reduce((accum, [propName, propType]) => {
+    const { argument, type } = propType;
+    const meta = propTypesMeta[propName];
 
-      // Replace type with meta type if a meta type exists
-      if (meta && (t.isIdentifier(meta) || t.isArrayExpression(meta))) {
-        valuePath.replaceWith(maybeRequiredNode(meta, isRequired));
-
-        return;
-      }
-
-      // Replace `PropTypes.number` with 'int' if meta type is missing
-      if (typeName === 'number') {
-        valuePath.replaceWith(
-          maybeRequiredNode(t.identifier('int'), isRequired)
-        );
-        return;
-      }
-
-      if (illegalTypes[typeName]) {
-        throw new Error(
-          `Invalid type '${typeName}' for prop '${propName}'.\n${
-            illegalTypes[typeName]
-          }`
-        );
-      }
-    },
-    CallExpression(path) {
-      const parent = path.findParent(parent => parent.isObjectProperty());
-      const propName = parent && parent.node.key.name;
-      const callee = path.get('callee');
-      const calleeName = callee.isMemberExpression()
-        ? callee.node.object.name
-        : callee.node.name;
-      const argument = path.node.arguments[0];
-
-      switch (calleeName) {
-        case 'shape':
-          path.replaceWith(argument);
-          break;
-        case 'arrayOf':
-          path.replaceWith(t.arrayExpression([argument]));
-          break;
-        default:
-          throw new Error(
-            `Invalid function call '${calleeName}' in '${propName}'`
-          );
-      }
+    if (typesToStrip.includes(type) || (meta && meta.type === 'exclude')) {
+      return accum;
     }
-  });
+
+    if (type === 'shape' && meta) {
+      argument;
+      return {
+        ...accum,
+        [propName]: {
+          ...propType,
+          argument: transformPropTypes(propType.argument, meta.argument)
+        }
+      };
+    }
+
+    if (meta) {
+      return { ...accum, [propName]: { ...propType, ...meta } };
+    }
+
+    if (type === 'number') {
+      return { ...accum, [propName]: { ...propType, type: 'int' } };
+    }
+
+    if (illegalTypes[type]) {
+      throw new Error(
+        `Invalid type '${type}' for prop '${propName}'.\n${illegalTypes[type]}`
+      );
+    }
+
+    return { ...accum, [propName]: propType };
+  }, {});
 };
+
+module.exports = transformPropTypes;
